Refactor typedefs and resolvers
