+++ packages/gql/db.ts
@@ -1,0 +1,14 @@
+import { Pool } from 'pg'
+
+const db = new Pool({
+  host: process.env.POSTGRES_HOST || 'localhost',
+  port: (process.env.POSTGRES_PORT || 5432) as number,
+  database: process.env.POSTGRES_DATABASE || 'user',
+  user: process.env.POSTGRES_USER || 'user',
+  password: process.env.POSTGRES_PASSWORD || 'password',
+  max: 10,
+  idleTimeoutMillis: 60_000,
+  connectionTimeoutMillis: 5_000,
+})
+
+export default db
