+++ packages/gql/package.json
@@ -8,20 +8,24 @@
     "production": "ts-node --transpile-only index.ts"
   },
   "dependencies": {
-    "cross-env": "5.0.5",
-    "lib": "^1.0.0",
     "@types/express": "^4.17.17",
     "@types/node": "^20.4.10",
     "@types/pg": "^8.10.2",
     "@types/uuid": "^9.0.2",
     "apollo-server-express": "^3.12.0",
     "concurrently": "^8.2.0",
+    "cross-env": "5.0.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "graphql": "^16.7.1",
+    "lib": "^1.0.0",
     "nodemon": "^3.0.1",
     "pg": "^8.11.2",
+    "redis-info": "^3.1.0",
     "ts-node": "^10.9.1",
     "typescript": "^5.1.6"
+  },
+  "devDependencies": {
+    "@types/redis-info": "^3.0.0"
   }
 }
