+++ packages/gql/resolvers/bananas.ts
@@ -1,0 +1,2 @@
+
+export default () => '🍌'.repeat(1 + Math.floor(Math.random() * 32))
