+++ packages/gql/resolvers/index.ts
@@ -1,0 +1,13 @@
+import bananas from './bananas'
+import latestBlocks from './latestBlocks'
+import vaults from './vaults'
+import monitor from './monitor'
+
+export default {
+  Query: {
+    bananas,
+    latestBlocks,
+    vaults,
+    monitor
+  }
+}
