+++ packages/gql/typeDefs/latestBlock.ts
@@ -1,0 +1,9 @@
+import { gql } from 'apollo-server-express'
+
+export default gql`
+  type LatestBlock {
+    chainId: Int!
+    blockNumber: BigInt!
+    blockTimestamp: BigInt!
+  }
+`
