+++ packages/ingest/block/poller.ts
@@ -17,8 +17,7 @@ export default class BlockPoller implements Processor {
         await this.queue?.add(mq.q.__noJobName, {
           chainId: rpc.chain?.id,
           blockNumber: block.number.toString(),
-          blockTimestamp: block.timestamp.toString(),
-          queueTimestamp: (Math.round(Date.now() / 1000)).toString()
+          blockTimestamp: block.timestamp.toString()
         } as types.LatestBlock, {
           jobId: `${rpc.chain?.id}-${block.number}`,
         })
