+++ packages/lib/index.ts
@@ -1,5 +1,4 @@
 export { chains } from './chains'
-export { Monitor } from './monitor'
 export * as mq from './mq'
 export * as types from './types'
 export * as blocks from './blocks'
