+++ packages/lib/package.json
@@ -7,11 +7,9 @@
     "test": "mocha"
   },
   "dependencies": {
-    "@types/redis-info": "^3.0.0",
     "bullmq": "^4.7.3",
     "cross-env": "^7.0.3",
     "dotenv": "^16.3.1",
-    "redis-info": "^3.0.8",
     "ts-node": "^10.9.1",
     "typescript": "^5.1.6",
     "viem": "1.6.3"
